urgent commit

+++ app/src/main/java/com/example/apolusov/kotlintest/PointD.kt
@@ -1,9 +1,9 @@
 package com.example.apolusov.kotlintest
 
 data class PointD(
-    val x: Float,
-    val y: Float,
-    val text: Float
+    val x: Int,
+    val y: Int,
+    val text: Int
 ) {
     override fun toString() = "[$x, $y, $text]"
 }
